+++ Assets/MixedRealityToolkit.SDK/Features/UX/Prefabs/Pointers/DefaultMousePointer.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 840df22ff840d5540b58ca6088afde25
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
