+++ Assets/MixedRealityToolkit.SDK/Features/UX/Scripts/Pointers/DefaultMousePointer.cs
@@ -1,0 +1,212 @@
+﻿// Copyright (c) Microsoft Corporation. All rights reserved.
+// Licensed under the MIT License. See LICENSE in the project root for license information.
+
+using Microsoft.MixedReality.Toolkit.Physics;
+using Microsoft.MixedReality.Toolkit.Utilities;
+using UnityEngine;
+
+namespace Microsoft.MixedReality.Toolkit.Input
+{
+    /// <summary>
+    /// Internal Touch Pointer Implementation.
+    /// </summary>
+    public class DefaultMousePointer : BaseControllerPointer, IMixedRealityMousePointer
+    {
+        protected float timeoutTimer = 0.0f;
+
+        private bool isInteractionEnabled = false;
+
+        private bool cursorWasDisabledOnDown = false;
+
+        protected bool isDisabled = true;
+
+        #region IMixedRealityMousePointer Implementation
+
+        [SerializeField]
+        [Tooltip("Should the mouse cursor be hidden when no active input is received?")]
+        private bool hideCursorWhenInactive = true;
+
+        /// <inheritdoc />
+        public bool HideCursorWhenInactive => hideCursorWhenInactive;
+
+        [SerializeField]
+        [Range(0.01f, 1f)]
+        [Tooltip("What is the movement threshold to reach before un-hiding mouse cursor?")]
+        private float movementThresholdToUnHide = 0.1f;
+
+        /// <inheritdoc />
+        public float MovementThresholdToUnHide => movementThresholdToUnHide;
+
+        [SerializeField]
+        [Range(0f, 10f)]
+        [Tooltip("How long should it take before the mouse cursor is hidden?")]
+        private float hideTimeout = 3.0f;
+
+        /// <inheritdoc />
+        public float HideTimeout => hideTimeout;
+
+        #endregion IMixedRealityMousePointer Implementation
+
+        #region IMixedRealityPointer Implementation
+
+        /// <inheritdoc />
+        public override bool IsInteractionEnabled => isInteractionEnabled;
+
+        private IMixedRealityController controller;
+
+        /// <inheritdoc />
+        public override IMixedRealityController Controller
+        {
+            get { return controller; }
+            set
+            {
+                controller = value;
+                TrackingState = TrackingState.NotApplicable;
+
+                if (controller != null && gameObject != null)
+                {
+                    InputSourceParent = controller.InputSource;
+                    Handedness = controller.ControllerHandedness;
+                    gameObject.name = "Spatial Mouse Pointer";
+                }
+            }
+        }
+
+        /// <inheritdoc />
+        public override void OnPreSceneQuery()
+        {
+            // screenspace to ray conversion
+            transform.position = CameraCache.Main.transform.position;
+
+            Ray ray = new Ray(transform.position, transform.forward);
+            Rays[0].CopyRay(ray, float.PositiveInfinity);
+
+            if (MixedRealityRaycaster.DebugEnabled)
+            {
+                Debug.DrawRay(ray.origin, ray.direction * PointerExtent, Color.green);
+            }
+
+            // ray to worldspace conversion
+            gameObject.transform.position = transform.position + transform.forward * DefaultPointerExtent;
+        }
+
+        public override Vector3 Position
+        {       
+            get
+            {
+                return gameObject.transform.position;
+            }
+        }
+
+        #endregion IMixedRealityPointer Implementation
+
+        #region IMixedRealitySourcePoseHandler Implementation
+
+        /// <inheritdoc />
+        public override void OnSourceDetected(SourceStateEventData eventData)
+        {
+            if (RayStabilizer != null)
+            {
+                RayStabilizer = null;
+            }
+
+            base.OnSourceDetected(eventData);
+
+            if (eventData.SourceId == Controller?.InputSource.SourceId)
+            {
+                isInteractionEnabled = true;
+            }
+        }
+       
+
+        /// <inheritdoc />
+        public override void OnSourceLost(SourceStateEventData eventData)
+        {
+            base.OnSourceLost(eventData);
+
+            if (eventData.SourceId == Controller?.InputSource.SourceId)
+            {
+                isInteractionEnabled = false;
+            }
+        }
+
+        #endregion IMixedRealitySourcePoseHandler Implementation
+
+        #region IMixedRealityInputHandler Implementation
+
+        /// <inheritdoc />
+        public override void OnInputDown(InputEventData eventData)
+        {
+            cursorWasDisabledOnDown = isDisabled;
+
+            if (cursorWasDisabledOnDown)
+            {
+                SetVisibility(true);
+                transform.rotation = CameraCache.Main.transform.rotation;
+            }
+            else
+            {
+                base.OnInputDown(eventData);
+            }
+        }
+
+        /// <inheritdoc />
+        public override void OnInputUp(InputEventData eventData)
+        {
+            if (!isDisabled && !cursorWasDisabledOnDown)
+            {
+                base.OnInputUp(eventData);
+            }
+        }
+
+        public override void OnInputChanged(InputEventData<MixedRealityPose> eventData) { }
+
+        #endregion IMixedRealityInputHandler Implementation
+
+        #region MonoBehaviour Implementation
+
+        protected override void Start()
+        {
+            isDisabled = DisableCursorOnStart;
+
+            base.Start();
+
+            if (RayStabilizer != null)
+            {
+                RayStabilizer = null;
+            }
+
+            foreach (var inputSource in InputSystem.DetectedInputSources)
+            {
+                if (inputSource.SourceId == Controller.InputSource.SourceId)
+                {
+                    isInteractionEnabled = true;
+                    break;
+                }
+            }
+        }
+
+        private void Update()
+        {
+            if (!hideCursorWhenInactive || isDisabled) { return; }
+
+            timeoutTimer += Time.unscaledDeltaTime;
+
+            if (timeoutTimer >= hideTimeout)
+            {
+                timeoutTimer = 0.0f;
+                SetVisibility(false);
+                isDisabled = true;
+            }
+        }
+
+        protected virtual void SetVisibility(bool visible)
+        {
+            BaseCursor?.SetVisibility(visible);
+        }
+
+        #endregion MonoBehaviour Implementation
+
+
+    }
+}
