+++ Assets/MixedRealityToolkit.SDK/Features/UX/Scripts/Pointers/MousePointer.cs
@@ -3,7 +3,6 @@
 
 using Microsoft.MixedReality.Toolkit.Physics;
 using Microsoft.MixedReality.Toolkit.Utilities;
-using Microsoft.MixedReality.Toolkit.Input.UnityInput;
 using UnityEngine;
 
 namespace Microsoft.MixedReality.Toolkit.Input
@@ -11,50 +10,8 @@ namespace Microsoft.MixedReality.Toolkit.Input
     /// <summary>
     /// Internal Touch Pointer Implementation.
     /// </summary>
-    public class MousePointer : BaseControllerPointer, IMixedRealityMousePointer
+    public class MousePointer : DefaultMousePointer
     {
-        private float timeoutTimer = 0.0f;
-
-        private bool isInteractionEnabled = false;
-
-        private bool cursorWasDisabledOnDown = false;
-
-        private bool isDisabled = true;
-
-        #region IMixedRealityMousePointer Implementation
-
-        [SerializeField]
-        [Tooltip("Should the mouse cursor be hidden when no active input is received?")]
-        private bool hideCursorWhenInactive = true;
-
-        /// <inheritdoc />
-        bool IMixedRealityMousePointer.HideCursorWhenInactive => hideCursorWhenInactive;
-
-        [SerializeField]
-        [Range(0.01f, 1f)]
-        [Tooltip("What is the movement threshold to reach before un-hiding mouse cursor?")]
-        private float movementThresholdToUnHide = 0.1f;
-
-        /// <inheritdoc />
-        float IMixedRealityMousePointer.MovementThresholdToUnHide => movementThresholdToUnHide;
-
-        [SerializeField]
-        [Range(0f, 10f)]
-        [Tooltip("How long should it take before the mouse cursor is hidden?")]
-        private float hideTimeout = 3.0f;
-
-        /// <inheritdoc />
-        float IMixedRealityMousePointer.HideTimeout => hideTimeout;
-
-        #endregion IMixedRealityMousePointer Implementation
-
-        #region IMixedRealityPointer Implementation
-
-        /// <inheritdoc />
-        public override bool IsInteractionEnabled => isInteractionEnabled;
-
-        private IMixedRealityController controller;
-
         private MixedRealityMouseInputProfile mouseInputProfile = null;
 
         private MixedRealityMouseInputProfile MouseInputProfile
@@ -68,25 +25,6 @@ namespace Microsoft.MixedReality.Toolkit.Input
                     mouseInputProfile = mouseManager?.MouseInputProfile;
                 }
                 return mouseInputProfile;
-            }
-        }
-
-
-        /// <inheritdoc />
-        public override IMixedRealityController Controller
-        {
-            get { return controller; }
-            set
-            {
-                controller = value;
-                TrackingState = TrackingState.NotApplicable;
-
-                if (controller != null && gameObject != null)
-                {
-                    InputSourceParent = controller.InputSource;
-                    Handedness = controller.ControllerHandedness;
-                    gameObject.name = "Spatial Mouse Pointer";
-                }
             }
         }
 
@@ -108,76 +46,6 @@ namespace Microsoft.MixedReality.Toolkit.Input
             gameObject.transform.position = transform.position + transform.forward * DefaultPointerExtent;
         }
 
-        public override Vector3 Position
-        {       
-            get
-            {
-                return gameObject.transform.position;
-            }
-        }
-
-        #endregion IMixedRealityPointer Implementation
-
-        #region IMixedRealitySourcePoseHandler Implementation
-
-        /// <inheritdoc />
-        public override void OnSourceDetected(SourceStateEventData eventData)
-        {
-            if (RayStabilizer != null)
-            {
-                RayStabilizer = null;
-            }
-
-            base.OnSourceDetected(eventData);
-
-            if (eventData.SourceId == Controller?.InputSource.SourceId)
-            {
-                isInteractionEnabled = true;
-            }
-        }
-       
-
-        /// <inheritdoc />
-        public override void OnSourceLost(SourceStateEventData eventData)
-        {
-            base.OnSourceLost(eventData);
-
-            if (eventData.SourceId == Controller?.InputSource.SourceId)
-            {
-                isInteractionEnabled = false;
-            }
-        }
-
-        #endregion IMixedRealitySourcePoseHandler Implementation
-
-        #region IMixedRealityInputHandler Implementation
-
-        /// <inheritdoc />
-        public override void OnInputDown(InputEventData eventData)
-        {
-            cursorWasDisabledOnDown = isDisabled;
-
-            if (cursorWasDisabledOnDown)
-            {
-                BaseCursor?.SetVisibility(true);
-                transform.rotation = CameraCache.Main.transform.rotation;
-            }
-            else
-            {
-                base.OnInputDown(eventData);
-            }
-        }
-
-        /// <inheritdoc />
-        public override void OnInputUp(InputEventData eventData)
-        {
-            if (!isDisabled && !cursorWasDisabledOnDown)
-            {
-                base.OnInputUp(eventData);
-            }
-        }
-
-        /// <inheritdoc />
         public override void OnInputChanged(InputEventData<Vector2> eventData)
         {
             if (eventData.SourceId == Controller?.InputSource.SourceId)
@@ -195,67 +63,14 @@ namespace Microsoft.MixedReality.Toolkit.Input
                 }
             }
         }
-
-        public override void OnInputChanged(InputEventData<MixedRealityPose> eventData)
-        {
-            if (eventData.SourceId == Controller?.InputSource.SourceId)
-            {
-                if (UseSourcePoseData)
-                {
-                    UpdateMouseRotation(eventData.InputData.Rotation.eulerAngles);
-                }
-            }
-        }
-
-        #endregion IMixedRealityInputHandler Implementation
-
-        #region MonoBehaviour Implementation
-
-        protected override void Start()
-        {
-            isDisabled = DisableCursorOnStart;
-
-            base.Start();
-
-            if (RayStabilizer != null)
-            {
-                RayStabilizer = null;
-            }
-
-            foreach (var inputSource in InputSystem.DetectedInputSources)
-            {
-                if (inputSource.SourceId == Controller.InputSource.SourceId)
-                {
-                    isInteractionEnabled = true;
-                    break;
-                }
-            }
-        }
-
-        private void Update()
-        {
-            if (!hideCursorWhenInactive || isDisabled) { return; }
-
-            timeoutTimer += Time.unscaledDeltaTime;
-
-            if (timeoutTimer >= hideTimeout)
-            {
-                timeoutTimer = 0.0f;
-                BaseCursor?.SetVisibility(false);
-                isDisabled = true;
-            }
-        }
-
-        #endregion MonoBehaviour Implementation
-
         private void UpdateMouseRotation(Vector3 mouseDeltaRotation)
         {
-            if (mouseDeltaRotation.magnitude >= movementThresholdToUnHide)
-            { 
+            if (mouseDeltaRotation.magnitude >= MovementThresholdToUnHide)
+            {
                 if (isDisabled)
                 {
                     // if cursor was hidden reset to center
-                    BaseCursor?.SetVisibility(true);
+                    SetVisibility(true);
                     transform.rotation = CameraCache.Main.transform.rotation;
                 }
 
@@ -268,6 +83,26 @@ namespace Microsoft.MixedReality.Toolkit.Input
             }
 
             transform.Rotate(mouseDeltaRotation, Space.World);
+        }
+
+        protected override void Start()
+        {
+            base.Start();
+
+#if UNITY_EDITOR
+            if (UnityEditor.EditorWindow.focusedWindow != null)
+            {
+                UnityEditor.EditorWindow.focusedWindow.ShowNotification(new GUIContent("Press \"ESC\" to regain mouse control"));
+            }
+#endif
+
+            Cursor.visible = false;
+            Cursor.lockState = CursorLockMode.Locked;
+        }
+
+        protected override void SetVisibility(bool visible)
+        {
+            BaseCursor?.SetVisibility(visible);
         }
     }
 }
